Refactor, rm old theme
