+++ screen2.0/src/app/downloads/downloads.tsx
@@ -1,15 +1,12 @@
 "use client"
 
 import * as React from "react"
-import { Typography, Tabs, Tab, Box, Container, ThemeProvider } from "@mui/material"
-
+import { Tabs, Tab, Box, Container} from "@mui/material"
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2"
-
 import { QuickStart } from "./quick-start"
 import { DetailedElements } from "./detailed-elements"
 import { DataMatrices } from "./data-matrices"
-import { useMemo, useState } from "react"
-import { defaultTheme } from "../../common/lib/themes"
+import { useState } from "react"
 import { useRouter } from "next/navigation"
 import { ApolloQueryResult } from "@apollo/client"
 
@@ -52,26 +49,24 @@ export default function DownloadsPage(props: {
   }
 
   return (
-    // <ThemeProvider theme={defaultTheme}>
-      <Container>
-        <Grid2 mt={2} container spacing={2}>
-          <Grid2 xs={12}>
-            <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
-              <Tabs value={page} onChange={handleChange} aria-label="basic tabs example">
-                <Tab label="Quick Start" sx={{ textTransform: "none" }} {...a11yProps(0)} />
-                <Tab label="Detailed Elements" sx={{ textTransform: "none" }} {...a11yProps(1)} />
-                <Tab label="Data Matrices" sx={{ textTransform: "none" }} {...a11yProps(2)} />
-              </Tabs>
-            </Box>
-          </Grid2>
-          <Grid2 xs={12}>
-            <QuickStart value={page} biosamples={props.biosamples} />
-            <DetailedElements value={page} biosamples={props.biosamples} />
-            {/* Matrices being fed biosamples might be redundant */}
-            <DataMatrices value={page} biosamples={props.biosamples} matrices={props.matrices} searchParams={props.searchParams} />
-          </Grid2>
+    <Container>
+      <Grid2 mt={2} container spacing={2}>
+        <Grid2 xs={12}>
+          <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
+            <Tabs value={page} onChange={handleChange} aria-label="basic tabs example">
+              <Tab label="Quick Start" sx={{ textTransform: "none" }} {...a11yProps(0)} />
+              <Tab label="Detailed Elements" sx={{ textTransform: "none" }} {...a11yProps(1)} />
+              <Tab label="Data Matrices" sx={{ textTransform: "none" }} {...a11yProps(2)} />
+            </Tabs>
+          </Box>
         </Grid2>
-      </Container>
-    // </ThemeProvider>
+        <Grid2 xs={12}>
+          <QuickStart value={page} biosamples={props.biosamples} />
+          <DetailedElements value={page} biosamples={props.biosamples} />
+          {/* Matrices being fed biosamples might be redundant */}
+          <DataMatrices value={page} biosamples={props.biosamples} matrices={props.matrices} searchParams={props.searchParams} />
+        </Grid2>
+      </Grid2>
+    </Container>
   )
 }
