+++ screen2.0/src/common/theme-registry/theme-registry.tsx
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
-import NextAppDirEmotionCacheProvider from './EmotionCache';
+import NextAppDirEmotionCacheProvider from './emotion-cache';
 import theme from './theme';
 
 export default function ThemeRegistry({ children }: { children: React.ReactNode }) {
